Following changes were made to the last implementation: - Created an external port for tc_in - Created an external port for tc_out
The two additional ports should it allow to measure both signals and to determine the time in different hardware constellations of SpaceWire Time-Codes.

Because I wasn't sure at this moment whether the current status of the repository contained a valid implementation of the RouterPlatform, I decided to create a new branch for this purpose. As soon as the correct work of the implementation could be varified (e.g. by an hardware test) its necessary to merge this branch back into the master branch! (Because the changes were not so dramatic to justify a new branch)