Fatal error was resolved: Wrong assignment of the different clocks to rxclk resp. txclk. The valid settings are now set to: - sysfreq: 100 MHz - txclkfreq: 100 MHz - rx_impl: impl_fast - tx_impl: impl_fast - rxchunk: 2
Pins are connected as follows:
- clk & txclk are connected with 100 MHz clock
- rxclk is connected with 200 MHz clock

Default register values for ports are changed as follows:
- txdivcnt from 0x01 to 0x00 to allow a transfer rate of 100 Mbps (txclk is then divided by one)

Additional changes that were made:
- Input values to spwstream like 'Autostart' or 'Linkstart' are now taken from registers (same as txdivcnt)
This leads to the fact that ports of the router now do not try to establish a connection by themselves so the
value of the corresponding register must be changed manually. The advantage of this is that boards can physically be connected without any data transfer happening.